requirements.txt

colorama~= 0.4.6
numpy

tqdm~=4.65.0
six~=1.16.0

requests~=2.31.0
PyAudio~=0.2.13
pydub~=0.25.1
pyvts~=0.3.1
keyboard~=0.13.5
characterai~=0.6.0
python-dotenv~=1.0.0
elevenlabs~=0.2.19
humanize~=4.7.0




TTS==0.20.6
uvicorn~=0.23.2
fastapi
psutil~=5.9.5
voicevox-client
translate~=3.6.1
groq
PyPDF2
langchain
langchain-community
langchain-google-genai
google-generativeai
google-api-python-client
faiss-cpu
python-dotenv
openai
gtts
--index-url https://download.pytorch.org/whl/cu118
torch
torchvision
torchaudio 
git+https://github.com/openai/whisper.git@main
playwright